# apps/api/openapi.json
{
  "components": {
    "schemas": {
      "AddReviewersIn": {
        "properties": {
          "notify": {
            "default": false,
            "description": "True 면 새로 추가된 reviewer 의 등록 이메일로 검토 요청 메일을 best-effort 발송한다. SMTP 실패는 reviewer 생성을 되돌리지 않는다.",
            "title": "Notify",
            "type": "boolean"
          },
          "user_ids": {
            "items": {
              "type": "string"
            },
            "maxItems": 50,
            "title": "User Ids",
            "type": "array"
          }
        },
        "title": "AddReviewersIn",
        "type": "object"
      },
...
      "DecisionIn": {
        "properties": {
          "comment": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "status": {
            "description": "approved | rejected | changes_requested",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "DecisionIn",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "TransitionIn": {
        "properties": {
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "TransitionIn",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "MX 사업부 위키 백엔드. DocumentJSON v1.0 (`packages/shared/schemas/document.json`) 을 단일 진실 공급원(SSOT) 으로 사용한다.\n\n**핵심 흐름**: `POST /api/v1/documents` 한 번으로 본문 검증 → 섹션 번호 부여 → 위키링크/태그/용어집/감사로그/검색인덱스가 모두 자동 갱신된다.\n\n**Conditional write**: 본문 수정(PUT/PATCH) 은 `If-Match` 헤더 필수 (낙관적 동시성). GET 응답의 `ETag` 를 그대로 사용하면 된다.",
    "title": "MX White Paper API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/documents/{slug}/reviewers": {
      "get": {
        "operationId": "list_reviewers_api_v1_documents__slug__reviewers_get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response List Reviewers Api V1 Documents  Slug  Reviewers Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "리뷰어 목록 (reader+)",
        "tags": [
          "approvals"
        ]
      },
      "post": {
        "operationId": "add_reviewers_api_v1_documents__slug__reviewers_post",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddReviewersIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Add Reviewers Api V1 Documents  Slug  Reviewers Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "문서 리뷰어 추가 (editor+)",
        "tags": [
          "approvals"
        ]
      }
    },
    "/api/v1/documents/{slug}/reviewers/{user_id}": {
      "delete": {
        "operationId": "remove_reviewer_api_v1_documents__slug__reviewers__user_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "리뷰어 제거 (editor+)",
        "tags": [
          "approvals"
        ]
      }
    },
    "/api/v1/documents/{slug}/reviewers/{user_id}/decision": {
      "post": {
        "operationId": "submit_decision_api_v1_documents__slug__reviewers__user_id__decision_post",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisionIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Submit Decision Api V1 Documents  Slug  Reviewers  User Id  Decision Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "리뷰어 결정 제출 (본인만)",
        "tags": [
          "approvals"
        ]
      }
    },
...
    "/api/v1/documents/{slug}/transition": {
      "post": {
        "operationId": "transition_status_api_v1_documents__slug__transition_post",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransitionIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Transition Status Api V1 Documents  Slug  Transition Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "문서 상태 전이 (editor+/admin)",
        "tags": [
          "approvals"
        ]
      }
    },
...
  },
...
}